refactor: move convertAllKeysToSnakeCase to a new utility file and update its usage in useAppData hook
- Extracted convertAllKeysToSnakeCase function from utils.js to appDataCaseConversion.js for better organization.
- Updated useAppData hook to import the function from the new location.
- Removed redundant data query handling in useAppData to streamline the conversion process.
